requirements.txt
finnhub-python==2.4.20
yfinance==0.2.58
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.46.0
transformers==4.36.2
torch==2.2.0
spacy==3.7.2
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
tqdm==4.66.1